Move decorator to its own file
